requirements_clinica.txt
flask==2.3.3
flask-cors==4.0.0
flask-login==0.6.3
flask-sqlalchemy==3.0.5
werkzeug==2.3.7
cryptography==41.0.7
pyOpenSSL==23.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
flask-wtf==1.1.1
wtforms==3.0.1
email-validator==2.0.0 